backend/requirements.txt
Django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.8.6,<2.9.0
psycopg2-binary>=2.8.6,<2.9.0
drf-spectacular>=0.15.1,<0.16.0
django-cors-headers>=4.1.0,<4.2.0
django-rest-auth>=0.9.5,<1.0
djangorestframework-simplejwt>=5.2.2,<5.3.0
django-allauth>=0.54.0,<0.55
djangorestframework-camel-case>=1.4.2,<1.5.0

country-list>=1.0.0,<1.1.0

ptvsd>=4.3.2,<4.4.0

pdf2image>=1.16.3,<1.17.0
PyPDF2>=3.0.1,<3.1
pdfminer.six>=20221105,<20221106
pymupdf>=1.22.3,<1.23